+++ lib/design_raptor.dart
@@ -1,8 +1,9 @@
 library design_raptor;
 
 import 'dart:ui';
-
 import 'package:flutter/material.dart';
+import 'package:flutter/scheduler.dart';
 
 part 'views/raptorView.dart';
 part 'views/raptorLazy.dart';
+part 'views/raptorTouchAnimation.dart';