+++ lib/views/raptorView.dart
@@ -13,17 +13,15 @@ class RaptorButton extends StatefulWidget {
   final double? height, width;
   final Icon? icon;
   final Color? color;
-  final String title;
   final VoidCallback onSubmit;
-  final String? fontFamily;
+  final Text buttonText;
   RaptorButton(
       {this.height,
       this.width,
       this.icon,
       this.color,
-      required this.title,
       required this.onSubmit,
-      this.fontFamily});
+      required this.buttonText});
   @override
   _RapptorButtonState createState() => _RapptorButtonState();
 }
@@ -34,25 +32,21 @@ class _RapptorButtonState extends State<RaptorButton> {
     return InkWell(
       onTap: widget.onSubmit,
       child: Container(
-        height: widget.height ?? 50,
-        width: widget.width ?? 120,
+        height: widget.height ?? MediaQuery.of(context).size.height * 0.05,
+        width: widget.width ?? MediaQuery.of(context).size.width * 0.26,
         decoration: BoxDecoration(
             color: widget.color ?? Colors.blue,
             borderRadius: BorderRadius.circular(20)),
         child: Row(
+          crossAxisAlignment: CrossAxisAlignment.center,
           mainAxisAlignment: MainAxisAlignment.center,
           children: [
-            widget.icon ?? Icon(Icons.add),
-            SizedBox(
-              width: 20,
-            ),
-            Text(
-              widget.title,
-              style: TextStyle(
-                  fontSize: 20,
-                  fontWeight: FontWeight.w500,
-                  fontFamily: widget.fontFamily ?? ''),
-            )
+            widget.icon ??
+                Icon(
+                  Icons.add,
+                  size: 20,
+                ),
+            widget.buttonText
           ],
         ),
       ),
@@ -60,50 +54,191 @@ class _RapptorButtonState extends State<RaptorButton> {
   }
 }
 
-Future<bool> onWillPop() async {
-  return false;
+///
+///Animated Dialog box to display element on dialogbox
+///
+///[dialogText] used for display the text on dialog box
+///
+///[size] contains the size like height and width as per differnt device size
+///
+class RaptorDialog extends StatefulWidget {
+  final Text dialogText;
+  final Size size;
+  RaptorDialog({required this.dialogText, required this.size});
+  @override
+  State<StatefulWidget> createState() => RaptorDialogState();
 }
 
-///
-/// Raptor Dialog Box helps to display loading dialog with an indicator and message
-///
-/// User can't go back while dialog box is displaying on screen.
-raptorDialogView(
-    {required BuildContext context,
-    required String message,
-    String? fontFamily}) {
-  return showDialog(
-      barrierDismissible: false,
-      context: context,
-      builder: (ctx) {
-        return WillPopScope(
-            child: Container(
-                height: double.maxFinite,
-                width: double.maxFinite,
-                decoration: BoxDecoration(
-                  color: Colors.transparent,
-                ),
-                child: SimpleDialog(
-                    backgroundColor: Colors.grey.withOpacity(0.5),
-                    children: <Widget>[
-                      Column(
+class RaptorDialogState extends State<RaptorDialog>
+    with SingleTickerProviderStateMixin {
+  late AnimationController controller;
+  late Animation<double> scaleAnimation;
+  late Animation<double> height, width;
+
+  @override
+  void initState() {
+    super.initState();
+
+    controller =
+        AnimationController(vsync: this, duration: Duration(milliseconds: 450));
+    scaleAnimation =
+        CurvedAnimation(parent: controller, curve: Curves.elasticInOut);
+
+    height = TweenSequence(<TweenSequenceItem<double>>[
+      TweenSequenceItem<double>(
+          tween: Tween(begin: 10, end: widget.size.height * 0.25), weight: 50)
+    ]).animate(controller);
+
+    width = TweenSequence(<TweenSequenceItem<double>>[
+      TweenSequenceItem<double>(
+          tween: Tween(begin: 10, end: widget.size.width * 0.7), weight: 50)
+    ]).animate(controller);
+    controller.addListener(() {
+      setState(() {});
+    });
+
+    controller.forward();
+  }
+
+  @override
+  void dispose() {
+    controller.dispose();
+    super.dispose();
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return Center(
+      child: Material(
+        color: Colors.transparent,
+        child: WillPopScope(
+            onWillPop: onWillPop,
+            child: ScaleTransition(
+              scale: scaleAnimation,
+              child: Container(
+                height: height.value,
+                width: width.value,
+                decoration: ShapeDecoration(
+                    color: Colors.white,
+                    shape: RoundedRectangleBorder(
+                        borderRadius: BorderRadius.circular(20.0))),
+                child: height.value == widget.size.height * 0.25
+                    ? Column(
                         mainAxisAlignment: MainAxisAlignment.center,
                         children: [
-                          CircularProgressIndicator(),
+                          CircularProgressIndicator(
+                            valueColor:
+                                AlwaysStoppedAnimation<Color>(Colors.red),
+                          ),
                           SizedBox(
                             height: 20,
                           ),
-                          Text(
-                            message,
-                            style: TextStyle(
-                                fontSize: 16,
-                                fontWeight: FontWeight.w500,
-                                fontFamily: fontFamily ?? ''),
+                          Padding(
+                            padding: EdgeInsets.symmetric(horizontal: 20.0),
+                            child: Wrap(
+                              children: [widget.dialogText],
+                            ),
                           )
                         ],
                       )
-                    ])),
-            onWillPop: onWillPop);
-      });
+                    : Container(),
+              ),
+            )),
+      ),
+    );
+  }
+
+  Future<bool> onWillPop() async {
+    return false;
+  }
 }
 
+
+
+
+ // Container(
+            //     height: 200,
+            //     child: Hero(
+            //       tag: 'flutterLogo',
+            //       child: GestureDetector(
+            //         onTap: () {
+            //           Navigator.push(
+            //               context,
+            //               MaterialPageRoute(
+            //                   builder: (BuildContext context) => Details()));
+            //         },
+            //         child: FlutterLogo(
+            //           size: 100,
+            //         ),
+            //       ),
+            //     )),
+            // AnimatedContainer(
+            //     color: colorContainer,
+            //     height: height,
+            //     duration: Duration(seconds: 1),
+            //     child: TweenAnimationBuilder(
+            //       tween: Tween<double>(begin: 0, end: 1),
+            //       duration: Duration(seconds: 2),
+            //       child: ListView.builder(
+            //           itemCount: 10,
+            //           itemBuilder: (context, index) {
+            //             return Container(
+            //               height: 60,
+            //               child: Card(
+            //                 color: Colors.blue,
+            //                 child: ListTile(
+            //                   title: Text('$index'),
+            //                   trailing: Text("Price: 1200.00 Rs"),
+            //                   leading: AnimatedBuilder(
+            //                       animation: controller,
+            //                       builder: (BuildContext context, _) {
+            //                         return IconButton(
+            //                           onPressed: () {
+            //                             statusAnimation
+            //                                 ? controller.reverse()
+            //                                 : controller.forward();
+            //                             setState(() {});
+            //                           },
+            //                           icon: Icon(
+            //                             Icons.home,
+            //                             size: sizeAnimation.value,
+            //                             color: colorAnimation.value,
+            //                           ),
+            //                         );
+            //                       }),
+            //                 ),
+            //               ),
+            //             );
+            //           }),
+            //       builder: (context, double value, Widget child) {
+            //         return Opacity(
+            //           opacity: value,
+            //           child: Padding(
+            //             padding: EdgeInsets.only(top: value * 20),
+            //             child: child,
+            //           ),
+            //         );
+            //       },
+            //     )),
+
+            // Center(
+      //     child: RaptorButton(
+      //   onSubmit: () => raptorDialogView(
+      //       context: context,
+      //       dialogText: Text(
+      //         "please Wait",
+      //         style: TextStyle(fontSize: 20, color: Colors.blue),
+      //       )),
+      //   buttonText: Text("Demo"),
+      // )
+      //     child: RaptorLoading(
+      //   loadCount: 3,
+      //   shrinkWrap: true,
+      //   listDetail: data,
+      //   itemBuilder: (context, index) {
+      //     return ListTile(
+      //         trailing: Icon(Icons.time_to_leave),
+      //         title: Text((index + 1).toString()));
+      //   },
+      // )
+      // ),