+++ CHANGELOG.md
@@ -1,3 +1,10 @@
-## 0.0.1
+## 0.0.2
 
-* TODO: Describe initial release.
+* Raptor Button Text style support provided for text designing.
+* As Device size Raptor Button size fixed.
+* Raptor Dialog box modified with different device size fixed.
+* Scale transition animation support provided to RaptorDialog.
+* Animated RaptorTouch button feature implemented for Animated Icon Button.
+* bug fix
+
+